requirements.txt
# Airflow
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.7.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data processing - PySpark
pyspark==3.5.0
py4j==0.10.9.7

# Traditional data processing (for compatibility)
pandas==2.1.4
numpy==1.25.2

# HTTP requests
requests==2.31.0

# Configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-env==1.1.3
pytest-html==4.1.1
mock==5.1.0
coverage==7.3.2

# Code quality
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Documentation
sphinx==7.2.6

# Additional PySpark dependencies
findspark==2.0.1 